Move me controller to user controller
